finished up assignment

+++ db/migrate/20220914000234_add_price_to_books.rb
@@ -1,5 +1,5 @@
 class AddPriceToBooks < ActiveRecord::Migration[6.1]
   def change
-    add_column :books, :price, :integer
+    add_column :books, :price, :decimal
   end
 end
